requirements.txt
torch
torchaudio
transformers
peft
datasets
openai
google-generativeai
wandb
requests
google-api-python-client
IMDbPY
tqdm
python-dotenv
langdetect
scipy
vllm
tokenizers
sentencepiece
python-dotenv